split options.jl
